refactor: change directory for Spinner
